+++ priv/repo/migrations/20220430204415_exzeitable_add_pg_trgm.exs
@@ -1,0 +1,7 @@
+defmodule TestWeb.Repo.Migrations.ExzeitableAddPgTrgm do
+  use Ecto.Migration
+
+  def change do
+    execute("CREATE EXTENSION pg_trgm", "DROP EXTENSION pg_trgm")
+  end
+end